+++ starfish/tool/command/account_create_command.py
@@ -1,6 +1,7 @@
 """
 
-    Tool command 'Create' ( an account )
+    Tool command Account Create
+
 
 """
 import logging
@@ -11,13 +12,13 @@ from .command_base import CommandBase
 logger = logging.getLogger(__name__)
 
 
-class CreateAccountCommand(CommandBase):
+class AccountCreateCommand(CommandBase):
 
-    def __init__(self, subparser):
-        super().__init__('create', subparser)
+    def __init__(self, sub_parser=None):
+        super().__init__('create', sub_parser)
 
-    def create_parser(self, subparser):
-        parser = subparser.add_parser(
+    def create_parser(self, sub_parser):
+        parser = sub_parser.add_parser(
             self._name,
             description='Create a new account',
             help='Create a new account'
@@ -33,6 +34,7 @@ class CreateAccountCommand(CommandBase):
             nargs='?',
             help='Optional name of the keyfile to write the account key data to'
         )
+        return parser
 
     def execute(self, args, output):
         account = Account.create(args.password)