+++ starfish/tool/command/account_get_token_command.py
@@ -1,6 +1,6 @@
 """
 
-    Command Send Token
+    Command Account Get Token
 
 """
 import logging
@@ -14,17 +14,20 @@ from .command_base import CommandBase
 logger = logging.getLogger(__name__)
 
 
-class SendTokenCommand(CommandBase):
+DEFAULT_AMOUNT = 10
 
-    def __init__(self, subparser):
-        super().__init__('token', subparser)
+
+class AccountGetTokenCommand(CommandBase):
+
+    def __init__(self, sub_parser=None):
+        super().__init__('token', sub_parser)
 
     def create_parser(self, subparser):
 
         parser = subparser.add_parser(
             self._name,
-            description='Send tokens for an account address',
-            help='Send tokens from one account to the another'
+            description='Get tokens for an account address',
+            help='Get some token from the development network'
         )
 
         parser.add_argument(
@@ -43,14 +46,12 @@ class SendTokenCommand(CommandBase):
         )
 
         parser.add_argument(
-            'to_address',
-            help='Address to send data too',
-        )
-
-        parser.add_argument(
             'amount',
-            help=f'Amount to send'
+            nargs='?',
+            default=DEFAULT_AMOUNT,
+            help=f'Amount to request. Default {DEFAULT_AMOUNT}'
         )
+        return parser
 
     def execute(self, args, output):
         if not Web3.isAddress(args.address):
@@ -58,16 +59,12 @@ class SendTokenCommand(CommandBase):
             return
 
         account = Account(args.address, args.password, key_file=args.keyfile)
-        amount = float(args.amount)
-        to_address = args.to_address
 
         network = self.get_network(args.url)
-        logger.debug(f'sending tokens from account {account.address} to account {to_address}')
-        network.send_token(account, to_address, amount)
+        network.request_test_tokens(account, int(args.amount))
 
+        logger.debug(f'requesting tokens for account {account.address}')
         balance = network.get_token_balance(account)
-        output.add_line(f'Send {amount} tokens from account: {args.address} to account {to_address}')
+        output.add_line(f'Get {args.get_command} for account: {args.address} balance: {balance}')
         output.set_value('balance', balance)
-        output.set_value('from_address', args.address)
-        output.set_value('to_address', args.to_address)
-        output.set_value('amount', amount)
+        output.set_value('address', args.address)